scraper-jumbo/requirements.txt
# Dependencias para Scraper Jumbo
# Versión: 1.0

# HTTP requests
requests>=2.28.0

# HTML parsing
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Configuración
PyYAML>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Logging y utilidades
colorama>=0.4.6

# Selenium (opcional para contenido dinámico)
# selenium>=4.5.0
# webdriver-manager>=4.0.0

# Generación de documentación
Markdown>=3.4.0

# Desarrollo
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
